Add files via upload
This DQN is inspired by the DQN described in the Part 2 of this tutorial - https://medium.com/emergent-future/simple-reinforcement-learning-with-tensorflow-part-0-q-learning-with-tables-and-neural-networks-d195264329d0. I have however changed the single 1/0 decision output to a one hot vector, so that I can use the same network for more implementations by just changing the environment.